+++ compose-md/core/src/test/java/dev/slav/composemd/ui/component/MdComponentsVisitorUnsupportedNodesTest.kt
@@ -1,15 +1,15 @@
 package dev.slav.composemd.ui.component
 
-import dev.slav.composemd.plugin.MdPlugin
+import dev.slav.composemd.ComposeMd
 import org.commonmark.node.Document
 import org.commonmark.node.Node
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.extension.ExtendWith
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.ParameterizedTest.ARGUMENTS_PLACEHOLDER
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
+import org.mockito.InjectMocks
 import org.mockito.Mock
 import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.kotlin.verifyNoInteractions
@@ -19,14 +19,10 @@ import org.mockito.kotlin.verifyNoInteractions
 class MdComponentsVisitorUnsupportedNodesTest {
 
     @Mock
-    lateinit var plugin: MdPlugin<Node>
+    lateinit var composeMd: ComposeMd
 
+    @InjectMocks
     lateinit var classUnderTest: MdComponentsVisitor
-
-    @BeforeEach
-    fun initVisitor() {
-        classUnderTest = MdComponentsVisitor(listOf(plugin))
-    }
 
     @ParameterizedTest(
         name = "GIVEN $ARGUMENTS_PLACEHOLDER, " +
@@ -37,7 +33,7 @@ class MdComponentsVisitorUnsupportedNodesTest {
     fun createdForParagraphByMatchingPlugin(node: Node) {
         node.accept(classUnderTest)
 
-        verifyNoInteractions(plugin)
+        verifyNoInteractions(composeMd)
     }
 
     companion object {